2_opencv/03_yolo/scripts/requirements.txt
ultralytics==8.3.74
opencv-contrib-python ==  4.10.0.84
opencv-python   ==  4.8.0.74
gradio==5.20.1
cvzone==1.6.1
pillow==11.1.0
numpy==1.24.3